update timestamp type

+++ src/main/java/com/pokerstar/api/domain/entity/admin/Admin.java
@@ -33,7 +33,7 @@ public class Admin implements Serializable {
 
     private BigDecimal admin_max_confirm_withdraw_amount;
 
-    private int admin_create_time;
+    private long admin_create_time;
 
     private String admin_remark;
 }
