+++ src/main/java/com/pokerstar/api/domain/entity/channel/Channel.java
@@ -34,9 +34,9 @@ public class Channel implements Serializable {
 
     private String channel_ip;
 
-    private int channel_create_time;
+    private long channel_create_time;
 
-    private int channel_last_update_time;
+    private long channel_last_update_time;
 
     private  String channel_remark;
 }
